+++ src/app/app-routing.module.ts
@@ -24,6 +24,7 @@ import { SystemComponent } from './system/system.component';
 import { UsersComponent } from './system/users/users.component';
 // user
 import { DashboardComponent } from './user/dashboard/dashboard.component';
+import { OverviewComponent } from './user/overview/overview.component';
 import { UserComponent } from './user/user.component';
 // search results and resource viewer
 import { ResourceComponent } from './workspace/resource/resource.component';
@@ -32,7 +33,7 @@ import { ResultsComponent } from './workspace/results/results.component';
 const routes: Routes = [
     {
         path: '',
-        component: MainComponent
+        component: OverviewComponent
     },
     {
         path: 'help',